First stab at a procmail-ready application of all this great code.
The dbm method doesn't work for me yet, but you can use it as-is
with the pickle method, just invoke it from procmail with the -f
option.

I had to make a minor change to the classifier so it would write back
modified values to the database.  I suppose I could have done this
by subclassing WordInfo's __setattr__ with a callback to the
containing PersistentGrahamBayes class, but this way is cleaner and
should incur no negligible penalty for the original GrahamBayes
class.  I hope this is okay with Tim :^)