central_limit_compute_population_stats()/ central_limit_compute_population_stats2()/ central_limit_compute_population_stats3():
Moved common code into new helper method _add_popstats().
Also arranged to add the new stats to any stats that may already
exist.  This allows a form of incremental training, so long as
you never try to forget a msg after one of the
compute_population_stats() methods is called.  It's unknown whether it
will be a *useful* form of incremental training -- it's cheating, but
there is no non-cheating way short of retraining on every msg ever
trained on.