A number of changes related to the folder selector dialog - code generally inspired by a couple of nice patches by Tony Meyer.  Should be no visible changes, but under the covers:
* worked out we were using short-term EIDs for folders, causing the
exchange server to fail.  So we are back to a faster MAPI version.
Deleted the other variants - as I said in the comments, CVS is your
friend <wink>.  I was kind enough leave an indication of the last CVS
revision with the old code though!

* Only build the folder structure as the folder is expanded.  Thus, the
entire folder hierarchy is no longer walked when displaying the dialog.
This speeds up the code considerable for exchange server users, where the
public folder hierarchy is both huge, and never needed by SpamBayes.

* We generally use the msgstore, rather than hitting MAPI directly,
abstracting away some ugly code.  Thus we are tied tighter to the
spambayes manager object.

* Drop all concepts of "default store" - all folder/message IDs are
expected to be a tuple of (store_id, item_id).  (This code was just
a hangover to prevent CVS users of spambayes from needing to redefine
all their folders when we first moved to the (store_id, item_id) scheme.