This is an initial crack at a Lotus Notes "integration" of Spambayes.  It is currently tested on Notes 5.0.11, and the term "tested" is used somewhat loosely.  This is definitely alpha level software, so backup early, backup often.
This is not a plugin as in the Outlook plugin.  It is a command line program
that interfaces with Notes, applying the Spambayes filter to mail in a Notes
mailbox.  It is working well for me at the moment, and I'm sure I'll find plenty
of things I want to enhance.  I have no plans at the moment for an
Outlookish style plugin for Notes.