Add a new filter option - "save_spam_info", default=True.  If not set, then we don't attempt to save the message before moving it (meaning spam score and original folder are not saved).
Option added mainly to see if it helps any of the "flagged as read" or
"leaves duplicate message" bugs we see - but note that in some cases the
message save would fail anyway, meaning this option was always implicit.

Fix some other incremental training assumptions re this spam score - we
assumed the score existed, but as mentioned above, this was always a
dangerous assumption.  Incremental training should still be effective
when the spam score can/is not be saved.

Updated the test suite to test all of this (which was a good thing as it
caught me out a couple of times!)