Remove an old comment.
Change the default (double-click) behaviour of the tray to "review messages" rather
than "display information".  It seems likely that this would be the most commonly
needed action.

Since 15/10/03, SetDefaultItem has been available for menus in win32all, so use that
as we should.  (It appears to set the font of the item correctly, but not have any effect
in terms of action, so still capture the double-click ourselves.  Someone correct me if
I've done this wrongly).