Clean out some cruft that isn't necessary with the latest version of twisted (1.1)
Add two new Message classes, one for messages that are stored in memory,
and one for messages that are re-generated each time the message is loaded.

Start our UIDs at 1, not 0, because Eudora likes this more.

Don't override the imported name "message".

Add a new folder - INBOX - this holds any messages from SpamBayes to the user.
(Having INBOX as an alias for Spam wasn't working well, and being able to communicate
within the confines of the mailer is nice, too).

Don't let the user set so much via the command line.  Use a config file, you lazy person.
(And I'll put in Skip's -o/--option flag at some point, so it can be done if really necessary).
(Not having the option there means that there is less trouble later on if -d/-D are changed
to -p and -d or something like that).

This actually appears to work with Eudora at the moment!