Richie hasn't managed to find time to test if this is causing a problem or not (and I'm not smart enough to <wink>), but he suspects that it could be, and given that this doesn't really break or change anything much at all, that's good enough for me, and fixing it before the release sounds like a good idea.
Basically, do the db save after each review page *before* writing anything to the
browser, so that if the user has (naughty naughty) closed the browser already,
it'll still do the save.  If they close it before the first message is saved, it'll still
try and write (the "training..." message), but that'd have to be very fast.