Use BODY.PEEK[] instead of RFC822.PEEK, which isn't really valid.
Report time taken a little less pedantically.

Fix semi-serious problem with handling unparseable messages.  What it was
doing was printing out an error and then returning from get_substance.  This
meant that the script kept going, but that the classifying/training was done on
an empty message - so when training the ham/spam count would just be incremented
and when classifying a blank message (with the sb headers) would be created
instead of the malformed one.

This now does what sb_server does - if the message is unparseable, then the
traceback is written out, and the program keeps going. When it's trained or
classified, mboxutils.get_message handily wraps the raw text and ignores the
headers.  When it's saved out we save it with an exception header.  (This
isn't *exactly* what sb_server does - it doesn't do any training/classifying - but
it's pretty close).

Fixes [ 737967 ] Malformed messages break pop3proxy (et al)