I find it hard to believe that this bug has never been seen before, when it would have been there for a number of releases.  Maybe I'm wrong, but testing seems to indicate that this is an actual bug.
To experience it (before this update) - best done with a fresh corpus
directory, rather than the one that has all your stuff in it:
  1.  Run sb_server with -o globals:verbose:True
  2.  Set your days to expiry to 0.
  3.  Receive a message via sb_server.
  4.  Train the message.
  5.  Telnet to localhost:110 (eg) and do USER username then QUIT
You'll see in the verbose info that the message that you just trained has been
*untrained*.  This is not good!

This fixes the problem by adding a 'flags' argument to the various corpus
observers (to my knowledge, this is only the storage.Trainer).  The only use
of the flag at the moment is to pass a "don't train" flag to the observers
(specifically removeMessage) when expiring messages, which then doesn't
train.  Messages that are removed in other ways (like regular use of the review
page) are still trained as appropriate.

I would really appreciate it if someone else could give this a wizz to confirm both
the bug and the fix.

(This arose at some point because of two things: adding in a expiry with every
USER command, so that long-running sb_servers still expired (the initial one
happens before the observers are attached, so doesn't hurt), and expiring the
ham/spam caches, so that users didn't end up with really big directories).