First stab at a troubleshooting guide for the sb_server binary.
This only has a couple of questions (I couldn't think of what was
common at the moment), but more can be added as time goes on.

Importantly, it does have information about how to find the log files,
how to get to the FAQ, and the online version of this file, and that
sort of thing.