Handle people using a storage type that isn't pickle/dbm (like one of the sql ones).
If we trigger the TypeError problem, try to do our best to put the message together
anyway (this is better than just leaving an X-SpamBayes-Exception header, which
sb_server/sb_imapfilter currently do.  This is what sb_filter currently does.  Ideally,
we should move to the new FeedParser in the email package, but we need to do
that in some way that provides 2.2 users with an option.  Maybe use it if it is there?

Bugfix candidate.