Add the "sb_culler.py" script from the SB wiki by Andrew Dalke.
The idea is that you run this script continuously and it clears your POP3
server of spam & viruses (storing the spam locally so that you can review it
if necessary).  You can leave this running when you go on vacation or
check mail via your phone or whatever, so that you don't need to handle
so much junk with limited means.

Something like this is requested fairly regularly, so it seems worth distributing
it rather than just having on the wiki.  This way it should stay updated, too,
and can be made more user friendly.

I asked Andrew about adding it to the dist and he's fine with that.

This is an exact copy of the version that's on the wiki.  It's stable (Andrew says
he's been running it almost continuously since he wrote it), but is slightly out
of date (works with the spambayes source that was current at the time).
As time allows, I'll update it to match current code and try to make it a bit
easier for non-programmers to use.