Allow makeMessage to take a content parameter, which is the initial content of the message.
Allow FileMessages to be created without any arguments, so that they
are compatible with emal.Parser.  Add in assert statements to ensure
that the required information is still entered.

Implement FileMessage.setPayload().  The parent version is now deprecated,
and this should also be avoided if possible (see the comment).

Update Factory classes to allow initial content.

Have testing setup use the Factory classes - this tests them and avoids using
setPayload.

The tests run, and sb_server still works, so I'm fairly confident that this hasn't broken
anything.  If it does, fix it or let me know and I will ASAP.