Fix three bugs found by Sjoerd Mullender:
imap.close() should be self.imap_server.close()

There were some instances of BadIMAPResponse instead of BadIMAPResponseError
(obviously I was confused about what I was calling it when I wrote them).

Need to pass the message id to insert_exception_header, and use self.rfc822_key
as the key.

And one in response to a problem from Aaron Konstam:

isinstance check is wrong, so will never be true, so literals in the folder list will never
be handled correctly.