Change [Classifier] x-use_bigrams to a normal, not experimental option.  i.e. it's now [Classifier] use_bigrams.
Note that this has almost no effect (the only one I can think of is that it won't be
listed on the Experimental options page in sb_server/sb_imapfilter).

The option is still False by default, and if you have "x-use_bigrams" in your config
file, it'll still (silently) work.  You can remove the x- if you like.