Fix docstring.
Stop using the web interface.  This is against the point of the module, and everything
 exception configuration is now provided via the IMAP server itself.

Fix bug in getHeaders where negation wouldn't work correctly.

Fix use of assert.

Remove code duplication in flags()

Fix loading of dynamic messages to correctly generate the headers, so that envelope
 works.

Change Factory and FileMessage to fit the new style.

Change the fake email addresses to the same format as the notate_to option (i.e. @spambayes.invalid)

Improve the "about" message to include the docstring.

Add a dynamic stats message.

Improve the dynamic status message to include everything that would normally be on
 the web interface.

Add a "train as spam" folder, to separate out training and classifying as spam.

Use the message.insert_exception_header utility function.

Use twisted.Application in the new style to avoid deprecation warnings.