Tidy up docstring.
Change MessageInfoBase's methods so that recording & retrieving a message are not
 private methods and are more clearly named.

Change so that the messageinfodb doesn't get created/opened on import, but rather
 through utility functions like those in spambayes.storage.

Change Stats.py to use the new methods rather than the old global.

Remove the asTokens function in favour of the existing tokenize function.

Fix the include_evidence header to check for *H* and *S* explicitly rather than any
 token starting with *.