Oops.  The minimum timer values were changed from 0 to 0.4 a little while back because the addin considers zero to be an invalid value.  Unfortunately, the change triggers the following assert in EditNumberProcessor.
"""
        # xxx - this wont be right if min <> 0 :(
        assert self.min_val == 0, "sue me"
"""

The assert aborts initialization of the Advanced tab in SpamBayes Manager and
leaves it in an invalid state, causing a nifty error message box if you even
try to leave the tab without editing anything.  Until we get a chance to
figure out how to make EditNumberProcessor work with non-zero minimum values,
I think it's best just to set the minimums back to avoid the error.